findColleg Test
