+++ LinkedIn/src/test/java/testHomePage/TestBusinessSolutionsLinks.java
@@ -1,20 +1,19 @@
-package TestHomePage;
+package testHomePage;
 
 import base.CommonAPI;
-import homePageLinks.GeneralLinks;
+import homePageLinks.BusinessSolutionsLinks;
 import org.openqa.selenium.support.PageFactory;
 import org.testng.annotations.Test;
 import reporting.ApplicationLog;
 import reporting.TestLogger;
 
-public class TestGeneralLinks extends CommonAPI {
+public class TestBusinessSolutionsLinks extends CommonAPI {
 
     @Test
-    public void generalHomePageLinks() throws Exception {
+    public void businessSolutions() throws Exception{
         ApplicationLog.epicLogger();
         TestLogger.log(getClass().getSimpleName() + ": " + convertToString(new Object(){}.getClass().getEnclosingMethod().getName()));
-        GeneralLinks generalHomePageLinks = PageFactory.initElements(driver, GeneralLinks.class);
-        generalHomePageLinks.clickGeneralButton();
+        BusinessSolutionsLinks businessSolutionsLinks = PageFactory.initElements(driver,BusinessSolutionsLinks.class);
+        businessSolutionsLinks.clickBusinessSolutionsButton();
     }
-
 }
