+++ LinkedIn/src/test/java/testHomePage/TestForgetPassWordButton.java
@@ -1,18 +1,16 @@
-package TestHomePage;
+package testHomePage;
 
 import base.CommonAPI;
-import org.testng.Assert;
 import org.testng.annotations.Test;
 import reporting.ApplicationLog;
 import reporting.TestLogger;
 
-public class TestHomePageTitle extends CommonAPI {
+public class TestForgetPassWordButton extends CommonAPI {
 
     @Test
-    public void homePageTitle(){
+    public void forgetPassWordButton(){
         ApplicationLog.epicLogger();
         TestLogger.log(getClass().getSimpleName() + ": " + convertToString(new Object(){}.getClass().getEnclosingMethod().getName()));
-        String title = driver.getTitle();
-        Assert.assertEquals(title,"LinkedIn: Log In or Sign Up");
+        clickOnXpath("//a[@class='link-forgot-password']");
     }
 }
