+++ LinkedIn/src/test/java/testHomePage/TestLinkedMemberDirectory.java
@@ -1,4 +1,4 @@
-package TestHomePage;
+package testHomePage;
 
 import base.CommonAPI;
 import homePageLinks.LinkedMemberDirectory;
@@ -6,7 +6,6 @@ import org.openqa.selenium.support.PageFactory;
 import org.testng.annotations.Test;
 import reporting.ApplicationLog;
 import reporting.TestLogger;
-import sun.rmi.rmic.iiop.PrintGenerator;
 
 public class TestLinkedMemberDirectory extends CommonAPI {
