+++ LinkedIn/src/test/resources/LinkedInTestRunner.xml
@@ -13,16 +13,17 @@
 
     <test name = "LinkedIn Home">
         <classes>
-            <class name="TestHomePage.TestHomePageTitle"/>
-            <class name="TestHomePage.TestBrowseLinkedInLinks"/>
-            <class name="TestHomePage.TestBusinessSolutionsLinks"/>
-            <class name="TestHomePage.TestDirectoriesLinks"/>
-            <class name="TestHomePage.TestFindACollegue"/>
-            <class name="TestHomePage.TestForgetPassWordButton"/>
-            <class name="TestHomePage.TestGeneralLinks"/>
-            <class name="TestHomePage.TestLanguageSelection"/>
-            <class name="TestHomePage.TestLinkedInC2109Links"/>
-            <class name="TestHomePage.TestLinkedMemberDirectory"/>
+            <!--<class name="testHomePage.TestHomePageTitle"/>-->
+            <!--<class name="testHomePage.TestBrowseLinkedInLinks"/>-->
+            <!--<class name="testHomePage.TestBusinessSolutionsLinks"/>-->
+            <!--<class name="testHomePage.TestDirectoriesLinks"/>-->
+            <!--<class name="testHomePage.TestFindACollegue"/>-->
+            <!--<class name="testHomePage.TestForgetPassWordButton"/>-->
+            <!--<class name="testHomePage.TestGeneralLinks"/>-->
+            <!--<class name="testHomePage.TestLanguageSelection"/>-->
+            <!--<class name="testHomePage.TestLinkedInC2109Links"/>-->
+            <!--<class name="testHomePage.TestLinkedMemberDirectory"/>-->
+            <class name="testOperation.TestFindCollegue"/>
 
         </classes>
     </test>
